Document Kotlin Coroutines Structured Concurrency
This commit documents the concept of structured concurrency in Kotlin coroutines.
It adds examples that demonstrate how `GlobalScope.launch` creates coroutines that are not bound to the scope (hence, no parent, and is independent) and requires explicit `join()` for completion.

It also documents the `join()` requirement to keep the program running until all `GlobalScope` coroutines finish.