Improve the example of OrphanGlobalScope where we demonstrate that a global scope is an independent coroutine without any parent. We have a global scope inside a runBlocking and the runBlocking will not wait for the globalScope. So, the runBlocking will finish before the globalScope starts or finishes.
Document and demonstrate independent coroutines with `GlobalScope.launch`.

This commit documents and demonstrates how `GlobalScope.launch` creates independent coroutines that are not children of the launching scope.
It refactors the existing code to include a function to launch the coroutines.
It includes a coroutine launched with `GlobalScope.launch`, demonstrating that it operates independently of the `runBlocking` scope and is not a child of it.
It removes the second `GlobalScope` coroutine because it was redundant. It was interfering and disturbing the actual purpose.
It adds log messages to understand the flow of coroutines.